requirements.txt
pytest
Django==2.2.13
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
requests==2.23.0
sentry-sdk==0.16.2
pytest-factoryboy==2.0.3